+++ spec/slack-ruby-bot/commands/aliases_spec.rb
@@ -1,6 +1,6 @@
 describe SlackRubyBot do
   def client
-    SlackRubyBot::Client.new aliases: %w(:emoji: alias каспаров)
+    SlackRubyBot::Client.new aliases: %w[:emoji: alias каспаров B0?.@(*^$]
   end
   it 'responds to emoji' do
     expect(message: ':emoji: hi').to respond_with_slack_message('Hi <@user>!')
@@ -10,5 +10,8 @@ describe SlackRubyBot do
   end
   it 'responds to a non-English alias' do
     expect(message: 'каспаров hi').to respond_with_slack_message('Hi <@user>!')
+  end
+  it 'responds to an alias with special characters' do
+    expect(message: 'B0?.@(*^$ hi').to respond_with_slack_message('Hi <@user>!')
   end
 end